Refactoring: classes rearrangements and remove of unnecessary code
