+++ ecosystem/daos/__init__.py
@@ -1,0 +1,3 @@
+"""DAOs for ecosystem."""
+
+from .jsondao import JsonDAO
