+++ ecosystem/models/tier.py
@@ -1,0 +1,9 @@
+"""Tier of ecosystem membership."""
+
+
+# pylint: disable=too-few-public-methods
+class Tier:
+    """Tiers of ecosystem membership."""
+    MAIN: str = "MAIN"
+    MEMBER: str = "MEMBER"
+    CANDIDATE: str = "CANDIDATE"
