+++ tests/resources/configured_python_repository/.pylintrc
@@ -377,7 +377,7 @@ ignore-imports=no
 ignore-signatures=no
 
 # Minimum lines number of a similarity.
-min-similarity-lines=10
+min-similarity-lines=4
 
 
 [BASIC]
